refactor: Utils 클래스 내부 메소드 분리
1. Utils 클래스에 있던 메소드들을 성격에 맞게 분리하여 utils 디렉토리에 파일별로 정리
2. 상수값을 정의한 constants.js도 util 디렉토리로 이동
