+++ src/utils/common.js
@@ -1,0 +1,5 @@
+const getAscending = (numbers) => {
+  return numbers.sort((a, b) => a - b);
+};
+
+module.exports = { getAscending };
